Salta error validacion - ok

+++ microservicios/demo/src/main/java/com/example/domains/services/ActorServiceImpl.java
@@ -40,6 +40,12 @@ public class ActorServiceImpl implements IActorService {
 
 	@Override
 	public Actor add(Actor item) throws DuplicateKeyException, InvalidDataException {
+		
+		if(item == null)
+			throw new InvalidDataException("Faltan los datos");
+		if(item.isInvalid())
+			throw new InvalidDataException("Errores de validacion: ");
+
 		if(getOne(item.getActorId()).isPresent())
 			throw new DuplicateKeyException();
 		return dao.save(item);
@@ -47,6 +53,11 @@ public class ActorServiceImpl implements IActorService {
 
 	@Override
 	public Actor modify(Actor item) throws NotFoundException,InvalidDataException {
+		if(item == null)
+			throw new InvalidDataException("Faltan los datos");
+		if(item.isInvalid())
+			throw new InvalidDataException("Errores de validacion: ");
+		
 		if(getOne(item.getActorId()).isEmpty())
 			throw new NotFoundException();
 		return dao.save(item);
@@ -60,8 +71,9 @@ public class ActorServiceImpl implements IActorService {
 	}
 	
 	@Override
-	public void delete(Actor item) {
-
+	public void delete(Actor item) throws InvalidDataException {
+		if(item == null)
+			throw new InvalidDataException("Faltan los datos");
 		dao.deleteById(item.getActorId());
 	}
 }
